Move out start/help commands
