+++ handlers/help.py
@@ -1,0 +1,115 @@
+from aiogram import Router
+from aiogram.types import Message
+from aiogram.filters import Command, CommandStart
+
+router = Router()
+
+# Start command and basic information
+@router.message(CommandStart())
+async def cmd_start(message: Message):
+    """Handler for /start command"""
+    user_name = message.from_user.first_name
+    
+    welcome_text = (
+        f"(っ◔◡◔)っ <b>Привет, {user_name}!</b> Вот что я умею~\n\n"
+        
+        f"<b>🎁 О боте</b>\n"
+        f"Я как волшебная коробка с записками-идеями о том, что можно делать вместе с друзьями! "
+        f"Создай свой пул активностей, добавь туда интересные идеи и пригласи друзей. "
+        f"Каждый может дополнить вашу общую коллекцию своими идеями, а когда захотите чем-то заняться - "
+        f"я случайным образом выберу одну из активностей! (つ✧ω✧)つ\n\n"
+        
+        f"<b>★.:*・°☆ Пошаговая инструкция ☆°・*:.★</b>\n\n"
+        
+        f"<b>Шаг 1: Создай свой пул активностей</b>\n"
+        f"Используй команду /create_pool\n"
+        f"Придумай название для своего пула (например, 'Выходные', 'Настолки', 'Фильмы')\n\n"
+        
+        f"<b>Шаг 2: Добавь активности в пул</b>\n"
+        f"Используй команду /add_activity\n"
+        f"Выбери свой пул и напиши, чем хотелось бы заняться (например, 'Поход в кино', 'Пикник в парке')\n\n"
+        
+        f"<b>Шаг 3: Пригласи друзей</b>\n"
+        f"Используй команду /invite\n"
+        f"Выбери пул, куда хочешь пригласить друзей, и получи код приглашения\n"
+        f"Отправь этот код друзьям - только ты как создатель пула можешь приглашать других!\n\n"
+        
+        f"<b>Шаг 4: Друзья присоединяются к пулу</b>\n"
+        f"Твои друзья используют команду /join_pool [код]\n"
+        f"Теперь они тоже могут добавлять свои идеи в общий пул!\n\n"
+        
+        f"<b>Шаг 5: Все вместе пополняете пул активностями</b>\n"
+        f"Каждый участник может использовать /add_activity\n"
+        f"Чем больше разных идей - тем интереснее!\n\n"
+        
+        f"<b>Шаг 6: Выбирайте случайную активность</b>\n"
+        f"Когда не можете решить, чем заняться - используйте /select\n"
+        f"Я выберу случайную активность с учетом истории и справедливости!\n\n"
+        
+        f"<b>❗ Важно: Система штрафов для справедливости</b>\n"
+        f"Когда выбирается чья-то активность, на её автора накладывается небольшой штраф. "
+        f"Это значит, что в следующий раз его активности будут выпадать реже. "
+        f"Так всем будет интереснее, и активности будут разнообразнее! (◕‿◕✿)\n"
+        f"Штрафы постепенно уменьшаются со временем, так что не переживай!\n\n"
+        
+        f"<b>★.:*・°☆ Все команды ☆°・*:.★</b>\n\n"
+        
+        f"🌟 <b>/start</b> - Начать работу с ботом и посмотреть краткий гайд\n\n"
+        
+        f"🌸 <b>/create_pool</b> - Создать новый пул активностей\n"
+        f"   Создай свою категорию с активностями! Ты будешь создателем и сможешь приглашать других.\n\n"
+        
+        f"🌸 <b>/join_pool [код]</b> - Присоединиться к пулу\n"
+        f"   Нужен код приглашения от создателя пула.\n\n"
+        
+        f"🌸 <b>/invite</b> - Пригласить в пул\n"
+        f"   Только создатель пула может приглашать участников!\n\n"
+        
+        f"🌸 <b>/exit_pool</b> - Выйти из пула\n"
+        f"   Покинуть пул, если больше не хочешь участвовать.\n\n"
+        
+        f"🌸 <b>/my_pools</b> - Показать твои пулы\n"
+        f"   Список всех пулов, в которых ты участвуешь.\n\n"
+        
+        f"🍡 <b>/add_activity</b> - Добавить активность\n"
+        f"   Добавь новую активность в пул для выбора.\n\n"
+        
+        f"🍡 <b>/remove_activity</b> - Удалить активность\n"
+        f"   Удалит активность из пула по её тексту.\n\n"
+
+        f"🍡 <b>/list_activities</b> - Список активностей\n"
+        f"   Покажет все активности в выбранном пуле.\n\n"
+        
+        f"🍡 <b>/select [номер_пула]</b> - Выбрать активность\n"
+        f"   Выберет случайную активность с учетом справедливого распределения.\n\n"
+        
+        f"📝 <b>/pool_info [название_пула]</b> - Информация о пуле\n"
+        f"   Подробная информация о пуле, участниках и активностях.\n\n"
+        
+        f"📝 <b>/penalties</b> - Информация о штрафах\n"
+        f"   Покажет текущие штрафы всех участников во всех твоих пулах.\n\n"
+        
+        f"Напиши /start в любой момент, чтобы вернуться к краткому гайду! (๑˃ᴗ˂)ﻭ"
+    )
+    
+    await message.answer(welcome_text, parse_mode="HTML")
+
+@router.message(Command("help"))
+async def cmd_help(message: Message):
+    help_text = (
+        "📚 <b>Справка по командам</b>\n\n"
+        "<b>Управление пулами</b>:\n"
+        "/create_pool - Создать новый пул активностей\n"
+        "/join_pool - Присоединиться к существующему пулу\n"
+        "/invite - Пригласить участника в пул\n"
+        "/exit_pool - Выйти из пула\n"
+        "/my_pools - Показать ваши пулы\n\n"
+        "<b>Управление активностями</b>:\n"
+        "/add_activity - Добавить активность в пул\n"
+        "/remove_activity - Удалить активность из пула\n"
+        "/select - Выбрать случайную активность\n\n"
+        "<b>Информация</b>:\n"
+        "/pool_info - Показать информацию о пуле\n"
+        "/help - Показать эту справку"
+    )
+    await message.answer(help_text, parse_mode="HTML")
